Refactoring google login
