+++ client/src/utils.ts
@@ -1,0 +1,7 @@
+const validateEmail = (email: string): boolean => {
+  var re = /\S+@\S+\.\S+/;
+  console.log("testing email");
+  return re.test(email);
+};
+
+export { validateEmail };
